Remove unecessary conversion to uint64 in sort option

+++ server/monitor_sort_opts.go
@@ -25,10 +25,10 @@ func (d ByCid) Len() int {
 	return len(d)
 }
 func (d ByCid) Swap(i, j int) {
-	d[uint64(i)], d[uint64(j)] = d[uint64(j)], d[uint64(i)]
+	d[i], d[j] = d[j], d[i]
 }
 func (d ByCid) Less(i, j int) bool {
-	return d[uint64(i)].Val.cid < d[uint64(j)].Val.cid
+	return d[i].Val.cid < d[j].Val.cid
 }
 
 type BySubs []Pair
@@ -37,10 +37,10 @@ func (d BySubs) Len() int {
 	return len(d)
 }
 func (d BySubs) Swap(i, j int) {
-	d[uint64(i)], d[uint64(j)] = d[uint64(j)], d[uint64(i)]
+	d[i], d[j] = d[j], d[i]
 }
 func (d BySubs) Less(i, j int) bool {
-	return d[uint64(i)].Val.subs.Count() < d[uint64(j)].Val.subs.Count()
+	return d[i].Val.subs.Count() < d[j].Val.subs.Count()
 }
 
 type ByOutMsgs []Pair
@@ -49,10 +49,10 @@ func (d ByOutMsgs) Len() int {
 	return len(d)
 }
 func (d ByOutMsgs) Swap(i, j int) {
-	d[uint64(i)], d[uint64(j)] = d[uint64(j)], d[uint64(i)]
+	d[i], d[j] = d[j], d[i]
 }
 func (d ByOutMsgs) Less(i, j int) bool {
-	return d[uint64(i)].Val.outMsgs < d[uint64(j)].Val.outMsgs
+	return d[i].Val.outMsgs < d[j].Val.outMsgs
 }
 
 type ByInMsgs []Pair
@@ -61,10 +61,10 @@ func (d ByInMsgs) Len() int {
 	return len(d)
 }
 func (d ByInMsgs) Swap(i, j int) {
-	d[uint64(i)], d[uint64(j)] = d[uint64(j)], d[uint64(i)]
+	d[i], d[j] = d[j], d[i]
 }
 func (d ByInMsgs) Less(i, j int) bool {
-	return d[uint64(i)].Val.inMsgs < d[uint64(j)].Val.inMsgs
+	return d[i].Val.inMsgs < d[j].Val.inMsgs
 }
 
 type ByOutBytes []Pair
@@ -73,10 +73,10 @@ func (d ByOutBytes) Len() int {
 	return len(d)
 }
 func (d ByOutBytes) Swap(i, j int) {
-	d[uint64(i)], d[uint64(j)] = d[uint64(j)], d[uint64(i)]
+	d[i], d[j] = d[j], d[i]
 }
 func (d ByOutBytes) Less(i, j int) bool {
-	return d[uint64(i)].Val.outBytes < d[uint64(j)].Val.outBytes
+	return d[i].Val.outBytes < d[j].Val.outBytes
 }
 
 type ByInBytes []Pair
@@ -85,8 +85,8 @@ func (d ByInBytes) Len() int {
 	return len(d)
 }
 func (d ByInBytes) Swap(i, j int) {
-	d[uint64(i)], d[uint64(j)] = d[uint64(j)], d[uint64(i)]
+	d[i], d[j] = d[j], d[i]
 }
 func (d ByInBytes) Less(i, j int) bool {
-	return d[uint64(i)].Val.inBytes < d[uint64(j)].Val.inBytes
+	return d[i].Val.inBytes < d[j].Val.inBytes
 }
